common: Move splitVlanIds into new package
The common subpackage will be home to those routines that the ovs plugin
may use no matter the interface type/backend that is configuring.

Start with splitVlanIds, which is used by pretty much all code execution
flows currently.

Also move SplitVlanIds tests from tests/cni/plugin_test.go into
pkg/common/common.go as they are essentially unit tests.
