+++ pkg/common/common.go
@@ -1,0 +1,67 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// This package collects routines that plugins might need to use no
+// matter the interface backend (veth, sriov vf, userspace...)
+package common
+
+import (
+	"errors"
+	"sort"
+
+	"github.com/k8snetworkplumbingwg/ovs-cni/pkg/types"
+)
+
+func SplitVlanIds(trunks []*types.Trunk) ([]uint, error) {
+	vlans := make(map[uint]bool)
+	for _, item := range trunks {
+		var minID uint = 0
+		var maxID uint = 0
+		if item.MinID != nil {
+			minID = *item.MinID
+			if minID > 4096 {
+				return nil, errors.New("incorrect trunk minID parameter")
+			}
+		}
+		if item.MaxID != nil {
+			maxID = *item.MaxID
+			if maxID > 4096 {
+				return nil, errors.New("incorrect trunk maxID parameter")
+			}
+			if maxID < minID {
+				return nil, errors.New("minID is greater than maxID in trunk parameter")
+			}
+		}
+		if minID > 0 && maxID > 0 {
+			for v := minID; v <= maxID; v++ {
+				vlans[v] = true
+			}
+		}
+		var id uint
+		if item.ID != nil {
+			id = *item.ID
+			if minID > 4096 {
+				return nil, errors.New("incorrect trunk id parameter")
+			}
+			vlans[id] = true
+		}
+	}
+	if len(vlans) == 0 {
+		return nil, errors.New("trunk parameter is misconfigured")
+	}
+	vlanIds := make([]uint, 0, len(vlans))
+	for k := range vlans {
+		vlanIds = append(vlanIds, k)
+	}
+	sort.Slice(vlanIds, func(i, j int) bool { return vlanIds[i] < vlanIds[j] })
+	return vlanIds, nil
+}
